Add generic parameter < P extends Pool< O, T > > to PoolObject
This will allow to access the derived Pool class in derived PoolObject
classes, e.g., a VertexWithFeatures can access the features fields of its
VertexWithFeaturesPool directly.

Also, enforce that PoolObject proxies cannot be pointed to other Pools
(different from the one that created them.) This was before not required
explicitly, but would have let to problems with FeatureValues associated
to a proxy. In other cases it would have worked though. Still, we never
used it, so we'll just enforce it consistently.