The UndoRedoStack has a save point.
It simply tracks whether there have been changes since the last time
setSavePoint() has been called.
The implementation is simplist: we store in the save point the index
in the undo stack when setSavePoint() has been called. Then in
isSavePoint() we check whether the index is still the same.
Calling undo() resets the save point, to avoid the possibility of
combining undos and edits have reaching the same stack index value.