Improve the performance Pool.createRef() and Pool.releaseRef()
The class Pool used a ConcurrentLinkedQueue to store unused references.
But ConcurrentLinkedQueue has some disadvantaged:
* It is slow when used by multiple threads.
* It creates a new object for each ConcurrentLinkedQueue.add(...) which
  can put some load on the garbage collector.
* Unused references are never garbage collected.

This commit introduces a ThreadLocalSoftReferencePool to store the unused
references, it is much faster when used with multiple threads. Slightly
faster. And it does not put such load on the garbage collector. It also
allows the garbage collector to free the unused references if needed.