Tiny change to fix UndoRedoStack.clear()
The Mastodon Git Collaboration Plugin benefits a lot from the ability to
reload the Mastodon Project from disk. This reload can actually be
implemented fairly easily. The ModelGraph listeners are muted during the
reload, and a "graphRebuilt" event is triggered after the reload from disk
operation. The graphRebuilt requires a clearing of the undo redo history.
But currently the clearing of the undo redo stack doesn't work properly.
Adding a spot after a graphRebuilt event fails and triggers an
ArrayIndexOutOfBounds exception.

This commit fixes that problem.